+++ mycodo/inputs/sht31.py
@@ -15,9 +15,10 @@ from mycodo.inputs.sensorutils import calculate_vapor_pressure_deficit
 from mycodo.utils.database import db_retrieve_table_daemon
 
 
-def constraints_pass_positive_value(value):
+def constraints_pass_positive_value(mod_input, value):
     """
     Check if the user input is acceptable
+    :param mod_input: SQL object with user-saved Input options
     :param value: float or int
     :return: tuple: (bool, list of strings)
     """
@@ -27,7 +28,8 @@ def constraints_pass_positive_value(value):
     if value <= 0:
         all_passed = False
         errors.append("Must be a positive value")
-    return all_passed, errors
+    return all_passed, errors, mod_input
+
 
 # Measurements
 measurements_dict = {